feat: add notationFromTokens and isNotationToken
Add new notationFromTokens method used to create a dot notation string
from the tokens array returned from tokenize.

Add new method isNotationToken which checks if the token provided is in
fact a valid token. This is more of a utility function.

Changed the return type on the tokenize function to Tokens which is an
extension of Array. The Tokens class contains a _notation property which
is a readonly string of the original notation used to generate the
tokens.